// src/app/api/models/create-stock-movement-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateStockMovementDto {
  productId: number;

  /**
   * Cantidad (positiva para in/out/usage, puede ser negativa para adjustment)
   */
  quantity: number;
  reason?: string;
  type: 'in' | 'out' | 'adjustment' | 'usage';
}